roleGeneration for onPrem users
Move roleGeneration in wca_base to make it available for onPrem users.
