+++ ansible_ai_connect/ai/api/model_pipelines/wca/pipelines_saas.py
@@ -12,10 +12,9 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 
-import json
 import logging
 from abc import ABCMeta
-from typing import TYPE_CHECKING, Generic, Optional, cast
+from typing import TYPE_CHECKING, Generic, Optional
 
 import backoff
 from django.apps import apps
@@ -33,7 +32,6 @@ from ansible_ai_connect.ai.api.model_pipelines.exceptions import (
     WcaKeyNotFound,
     WcaModelIdNotFound,
     WcaNoDefaultModelId,
-    WcaRequestIdCorrelationFailure,
     WcaTokenFailure,
 )
 from ansible_ai_connect.ai.api.model_pipelines.pipelines import (
@@ -70,11 +68,8 @@ from ansible_ai_connect.ai.api.model_pipelines.wca.pipelines_base import (
     WcaModelRequestException,
     WcaTokenRequestException,
     ibm_cloud_identity_token_hist,
-    wca_codegen_role_hist,
 )
 from ansible_ai_connect.ai.api.model_pipelines.wca.wca_utils import (
-    Context,
-    InferenceResponseChecks,
     TokenContext,
     TokenResponseChecks,
 )
@@ -350,78 +345,6 @@ class WCASaaSRoleGenerationPipeline(
 
     def __init__(self, config: WCASaaSConfiguration):
         super().__init__(config=config)
-
-    # This should be moved to the base WCA class when it becomes available on-prem
-    def invoke(self, params: RoleGenerationParameters) -> RoleGenerationResponse:
-        if not settings.ANSIBLE_AI_ENABLE_ROLE_GEN_ENDPOINT:
-            raise FeatureNotAvailable
-
-        request = params.request
-        name = params.name
-        text = params.text
-        create_outline = params.create_outline
-        outline = params.outline
-        model_id = params.model_id
-        generation_id = params.generation_id
-
-        api_key = self.get_api_key(request.user)
-        model_id = self.get_model_id(request.user, model_id)
-
-        headers = self.get_request_headers(api_key, generation_id)
-        data = {
-            "model_id": model_id,
-            "text": text,
-            "create_outline": create_outline,
-        }
-        if name:
-            data["name"] = name
-        if outline:
-            data["outline"] = outline
-
-        @backoff.on_exception(
-            backoff.expo,
-            Exception,
-            max_tries=self.retries + 1,
-            giveup=self.fatal_exception,
-            on_backoff=self.on_backoff_codegen_role,
-        )
-        @wca_codegen_role_hist.time()
-        def post_request():
-            return self.session.post(
-                f"{self.config.inference_url}/v1/wca/codegen/ansible/roles",
-                headers=headers,
-                json=data,
-                verify=self.config.verify_ssl,
-            )
-
-        result = post_request()
-
-        x_request_id = result.headers.get(WCA_REQUEST_ID_HEADER)
-        if generation_id and x_request_id:
-            # request/payload suggestion_id is a UUID not a string whereas
-            # HTTP headers are strings.
-            if x_request_id != str(generation_id):
-                raise WcaRequestIdCorrelationFailure(model_id=model_id, x_request_id=x_request_id)
-
-        context = Context(model_id, result, False)
-        InferenceResponseChecks().run_checks(context)
-        result.raise_for_status()
-
-        response = json.loads(result.text)
-
-        name = response["name"]
-        files = response["files"]
-        outline = response["outline"]
-        warnings = response["warnings"] if "warnings" in response else []
-
-        from ansible_ai_connect.ai.apps import AiConfig
-
-        ai_config = cast(AiConfig, apps.get_app_config("ai"))
-        if ansible_lint_caller := ai_config.get_ansible_lint_caller():
-            for file in files:
-                file["content"] = ansible_lint_caller.run_linter(file["content"])
-
-        return name, files, outline, warnings
 
     def self_test(self) -> HealthCheckSummary:
         return HealthCheckSummary(
